ref: extract people sort functions into hooks
