+++ src/features/people/components/PeopleTableHead.tsx
@@ -1,43 +1,9 @@
-import { Link, useSearchParams } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { sortableColumns } from './sortConfig';
+import usePeopleSortParams from '../hooks/usePeopleSortParams';
 
 export const PeopleTableHead = () => {
-  const [searchParams, setSearchParams] = useSearchParams();
-
-  const currentSort = searchParams.get('sort');
-  const currentOrder = searchParams.get('order');
-
-  const handleSort = (sortBy: string) => {
-    const newParams = new URLSearchParams(searchParams);
-
-    const isCurrentSort = currentSort === sortBy;
-
-    if (!isCurrentSort) {
-      newParams.set('sort', sortBy);
-    } else if (isCurrentSort && !currentOrder) {
-      newParams.append('order', 'desc');
-    } else {
-      newParams.delete('sort');
-      newParams.delete('order');
-    }
-
-    setSearchParams(newParams);
-  };
-
-  const getSortIcon = (columnTitle: string): string => {
-    const params = Object.fromEntries(searchParams.entries());
-    const isSorted = params.sort === columnTitle;
-
-    if (!isSorted) {
-      return 'fas fa-sort';
-    }
-
-    if (params.order === 'desc') {
-      return 'fas fa-sort-down';
-    }
-
-    return 'fas fa-sort-up';
-  };
+  const { handleSort, getSortIcon } = usePeopleSortParams();
 
   return (
     <thead>
